Refactor code for check payment status feature
